Proposed version 2.1. - Use Float32 image data instead of UInt16 - Use satMask instead of ccdInfo.satLevel to identify saturated pixels.
python/Constants.py
- Noted that satLevel is no longer used by PyGuide.

python/FindStars.py
- Added satMask argument to findStars.
- Modified to use Float32 image data instead of UInt16.

python/Centroid.py
- Added satMask argument to centroid and basicCentroid;
  they now ignore ccdInfo.satLevel.
- Modified to use Float32 omage data instead of UInt16.

tests/testStarShape.py
tests/testStarShapeLong.py
tests/timeCentroid.py
tests/testCentroidLong.py
tests/testCentroid.py
- Supply null satMask for PyGuide 2.1.

src/RadProfModule.c
src/RadProfModule.h
- Modified to use Float32 image data, not UInt16.

doFindStars.py
- Added support for satMask.
- Bug fix: modified to handle nSat=None on output.

+++ src/RadProfModule.h
@@ -37,7 +37,7 @@ void g_radAsymm_free(
 );
 long radAsymm(
 	long inLenI, long inLenJ,
-	UInt16 data[inLenI][inLenJ],
+	Float32 data[inLenI][inLenJ],
 	Bool mask[inLenI][inLenJ],
 	long iCtr, long jCtr,
 	long rad,
@@ -46,7 +46,7 @@ long radAsymm(
 );
 long radAsymmWeighted(
 	long inLenI, long inLenJ,
-	UInt16 data[inLenI][inLenJ],
+	Float32 data[inLenI][inLenJ],
 	Bool mask[inLenI][inLenJ],
 	long iCtr, long jCtr,
 	long rad,
@@ -58,7 +58,7 @@ long radAsymmWeighted(
 );
 long radProf(
 	long inLenI, long inLenJ,
-	UInt16 data[inLenI][inLenJ],
+	Float32 data[inLenI][inLenJ],
 	Bool mask[inLenI][inLenJ],
 	long iCtr, long jCtr,
 	long rad,
@@ -70,7 +70,7 @@ long radProf(
 );
 long radSqProf(
 	long inLenI, long inLenJ,
-	UInt16 data[inLenI][inLenJ],
+	Float32 data[inLenI][inLenJ],
 	Bool mask[inLenI][inLenJ],
 	long iCtr, long jCtr,
 	long rad,
